refactor(order): split checkout into feature and widget
