+++ src/features/order/checkout/index.js
@@ -1,0 +1,1 @@
+export { useCheckout } from './model/useCheckout';
