+++ src/widgets/order-checkout/index.js
@@ -1,0 +1,1 @@
+export { default } from './ui/OrderCheckout';
